spec: Add basic team propose/confirm specs
This will be it for now to make sure the changes made to fix
21c2f7ead8c48e6fb0daf57edd7f211566c2422c are actually good. This was
necessary because I messed up on it a few times.

A ton of functionality not being tested but that can come later. We are
currently in the situation of having to backfill a bunch of tests for
existing code.