spec: Add convenience to make a pm
This is a bit annoying since it spews a warning in cinch-test (@channel
uninitialized), should probably get that fixed.

The more annoying thing is that cinch-test seems to fire two threads for
these messages. This was causing the bot to fire the handler twice and
raise an exception on the second, since the game has no current_round
by the time the second thread is processed (and therefore we get a
NoMethodError by calling in_mission_phase? on nil). Will have to
investigate why that happens in cinch-test and possibly get fixed?